Adding more improvements to the UI and changes visibility in Managements
The things that were done in this commit, were to focus the interface, not allow the user to maximize or minimize, and also not allow to change the size of it.
Also another thing that was done was to change the visibility of two functions of two handlers. One in each handler