Name improvements and modifications
In particular what was done was to change the names of the enum since they did not follow Clean Code. The business rules of the application domain were also separated and the Mocked Date class was moved to the Test package. The utilities class was also created to eliminate the duplication of code that was in the logic